Moving `ScheduleServiceActivation` to its right place
